bulleye_backend/requirements.txt
# Core Django

Django==4.2.11
djangorestframework==3.15.1

# Auth & Security

djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1

# Database (PostgreSQL)

psycopg2-binary==2.9.9

# Celery & Redis (Async tasks)

celery==5.4.0
redis==5.0.4
django-celery-beat==2.6.0
amqp>=5.2.0

# CORS (for frontend integration)

django-cors-headers==4.3.1

# API Utilities

requests==2.32.3

# Data Processing / Analytics

pandas==2.2.2
numpy==1.26.4

# Technical Analysis (optional but useful)

ta==0.11.0

# AI / LLM Integration

openai==1.35.10

# Web Server (Production)

gunicorn==22.0.0

# Optional: Caching

django-redis==5.4.0

# Optional: Environment config

dj-database-url==2.1.0

# Optional: WebSockets (future real-time alerts)

channels==4.1.0

# Optional: Image / media handling

Pillow==10.3.0
